// src/app/business-logic/model/tasks/artifact.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { ArtifactModeEnum } from './artifactModeEnum';
import { ArtifactTypeData } from './artifactTypeData';


export interface Artifact {
    /**
     * Entry key
     */
    key: string;
    /**
     * System defined type
     */
    type: string;
    mode?: ArtifactModeEnum;
    /**
     * Raw data location
     */
    uri?: string;
    /**
     * Raw data length in bytes
     */
    content_size?: number;
    /**
     * Hash of entire raw data
     */
    hash?: string;
    /**
     * Epoch time when artifact was created
     */
    timestamp?: number;
    type_data?: ArtifactTypeData;
    /**
     * User-defined list of key/value pairs, sorted
     */
    display_data?: Array<Array<string>>;
}