// src/app/business-logic/model/tasks/taskModelItem.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.14
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {IModelInfo} from '../../../webapp-common/experiments/shared/common-experiment-model.model';


export interface TaskModelItem {
    /**
     * The task model name
     */
    name: string;
    /**
     * The model ID
     */
    model: IModelInfo;
}